Restore sensibility in split controller
