+++ src/holosoma_inference/holosoma_inference_service/holosoma_service/holosoma_service/unitree_control/unitree_split_controller.py
@@ -1,29 +1,21 @@
-"""Split-body Unitree controller (arm_sdk + loco).
-
-``ControllerNode`` is the explicit node: it owns the SDK clients, the
-``ExoCommandSource`` target provider, and the command/heartbeat publishers, and
-runs the control loop. Each tick reads the newest ExoskeletonCmd, sends the
-14-DoF arm target ([left(7), right(7)]) and the base Twist, republishes the
-commanded arm target on ``/holosoma_cmd``, and beats a Heartbeat at 5 Hz.
-
-Each SDK client runs in its own spawned process (separate DDS). Loco enters
-FSM-501 BEFORE arm init, else arm_sdk is ignored.
-"""
+"""Split-body Unitree controller (arm_sdk + loco)."""
 
 from __future__ import annotations
 
 import numpy as np
 import rclpy
 import tyro
+from holosoma_input_msgs.msg import ExoskeletonCmd
 from holosoma_state_msgs.msg import Heartbeat
 from loguru import logger
 from rclpy.node import Node
+from rclpy.qos import QoSProfile, ReliabilityPolicy
 from sensor_msgs.msg import JointState
 
 from holosoma_inference.sdk.unitree_high_level import make_mp_arm_client, make_mp_loco_client
 from holosoma_inference.utils.rate import RateLimiter
-from holosoma_service.unitree_control.exo_command_source import ExoCommandSource
 
+EXOSKELETON_TOPIC = "/holosoma/exoskeleton_command"
 CMD_TOPIC = "/holosoma_cmd"
 HEARTBEAT_TOPIC = "/holosoma/heartbeat"
 CONTROL_RATE_HZ = 50.0
@@ -34,22 +26,24 @@ _ARM_JOINTS = ("shoulder_pitch", "shoulder_roll", "shoulder_yaw", "elbow", "wris
 ARM_JOINT_NAMES = [f"{side}_{j}" for side in ("left", "right") for j in _ARM_JOINTS]
 
 
-class ControllerNode(Node):
+class UnitreeSplitControllerNode(Node):
     def __init__(self, arm, loco) -> None:
-        super().__init__("holosoma_controller")
+        super().__init__("unitree_split_controller")
         self._arm = arm
         self._loco = loco
-        self._source = ExoCommandSource()
+        self._latest: ExoskeletonCmd | None = None
+        qos = QoSProfile(depth=1, reliability=ReliabilityPolicy.BEST_EFFORT)
+        self.create_subscription(ExoskeletonCmd, EXOSKELETON_TOPIC, self._cb, qos)
         self._cmd_pub = self.create_publisher(JointState, CMD_TOPIC, 10)
         self._hb_pub = self.create_publisher(Heartbeat, HEARTBEAT_TOPIC, 10)
         self._rate = RateLimiter(CONTROL_RATE_HZ)
 
     def run(self) -> None:
-        self._source.start()
         logger.info(f"control loop @ {CONTROL_RATE_HZ:.0f} Hz")
         tick = 0
         while rclpy.ok():
-            cmd = self._source.get_latest()
+            rclpy.spin_once(self, timeout_sec=0.0)  # drain pending callbacks (newest cmd)
+            cmd = self._latest
             if cmd is not None:
                 q_target = np.concatenate([cmd.q_left_arm, cmd.q_right_arm])
                 if self._arm is not None:
@@ -63,9 +57,8 @@ class ControllerNode(Node):
             tick += 1
             self._rate.sleep()
 
-    def stop(self) -> None:
-        self._source.stop()
-        self.destroy_node()
+    def _cb(self, msg: ExoskeletonCmd) -> None:
+        self._latest = msg
 
     def _publish_joint_command(self, q_target: np.ndarray) -> None:
         msg = JointState()
@@ -96,14 +89,14 @@ def main(iface: str = "eth0", no_arms: bool = False, no_loco: bool = False) -> N
         arm.speed_gradual_max()
 
     rclpy.init()
-    controller = ControllerNode(arm=arm, loco=loco)
+    controller = UnitreeSplitControllerNode(arm=arm, loco=loco)
     try:
         controller.run()
     except KeyboardInterrupt:
         pass
     finally:
         logger.info("shutting down …")
-        controller.stop()
+        controller.destroy_node()
         rclpy.shutdown()
         if loco is not None:
             loco.close()  # type: ignore[attr-defined]
