+++ src/holosoma_inference/holosoma_inference_service/holosoma_service/holosoma_service/unitree_control/wasd_controller_node.py
@@ -27,7 +27,7 @@ from geometry_msgs.msg import Twist
 from holosoma_input_msgs.msg import ExoskeletonCmd
 from rclpy.node import Node
 
-from holosoma_service.unitree_control.exo_command_source import EXOSKELETON_TOPIC
+from holosoma_service.unitree_control.unitree_split_controller import EXOSKELETON_TOPIC
 
 LIN_STEP = 0.1  # m/s per keypress
 ANG_STEP = 0.2  # rad/s per keypress
