+++ src/holosoma_inference/holosoma_inference_service/holosoma_service/setup.py
@@ -23,7 +23,7 @@ setup(
         "console_scripts": [
             "holosoma_node = holosoma_service.policy_control.holosoma_node:main",
             "retargeter_node = holosoma_service.retargetting.retargeter_node:main",
-            "controller_node = holosoma_service.unitree_control.controller_node:_cli",
+            "unitree_split_controller = holosoma_service.unitree_control.unitree_split_controller:_cli",
             "wasd_controller_node = holosoma_service.unitree_control.wasd_controller_node:main",
         ],
     },
